+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
-## [Unreleased](https://github.com/akhx/bitrix-optimization/compare/v1.0.0...HEAD)
+## [Unreleased](https://github.com/akhx/bitrix-optimization/compare/v1.1.0...HEAD)
+
+## [1.1.0](https://github.com/akhx/bitrix-optimization/compare/v1.0.0...v1.1.0) - 2021-07-07
 ### Добавлено
 - Блокировка конвертации по stop_convert=Y
-- Сброс ковертации по clear_cache=Y
+- Сброс ковертации по clear_cache=Y (isAdmin)
+- Для отладки LazyContent можно загружать сразу по include_lazy_content=Y (isAdmin)
 
 ## [1.0.0](https://github.com/akhx/bitrix-optimization/compare/v0.2.2...v1.0.0) - 2021-06-29
 ### Добавлено